requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
pdfplumber>=0.11.0
python-docx>=1.1.2
requests>=2.32.0
pytest>=8.0.0
